refactor: migrate ScrollSpy script to TypeScript and update Head component references
