+++ astro/package.json
@@ -33,6 +33,7 @@
 		"astro-link-validator": "github:rodgtr1/astro-link-validator",
 		"astro-mermaid": "^2.0.1",
 		"codemirror": "^6.0.2",
+		"es-toolkit": "1.46.1",
 		"jsonpath-plus": "^10.4.0",
 		"marked": "^17.0.3",
 		"marked-plaintify": "^1.1.1",
