+++ astro/src/components/Head.astro
@@ -60,7 +60,7 @@ const slug = Astro.params.slug;
   <link rel="sitemap" href="/sitemap-index.xml">
   <Font cssVariable="--font-inter-var" />
   <script defer src="/js/Search-0.2.3.js" type="text/javascript"></script>
-  <script src="../scripts/ScrollSpy.js"></script>
+  <script src="../scripts/ScrollSpy.ts"></script>
   <script src="../scripts/Visibility.ts"></script>
   {applyTheme && <script is:inline>
     const getTheme = () => {
